# Make Cover Theme 3x faster