Listening to `PREFILL_MEDIA_PRODUCT_TRANSITION` instead of `TIME_UPDATE` is about 3x faster, because it fires when the user clicks on a song, rather than when the song actually starts playing.

|     | Cached | Uncached |
|-----|--------|----------|
| New | 150ms  | 450ms    |
| Old | 600ms  | 1200ms   |